Former-commit-id: 599a597f6dd8d62e2c1e68805d3c344bd7c88173

+++ Programs/TestBed/Classes/Tests/FullscreenTest.cpp
@@ -113,7 +113,7 @@ void FullscreenTest::LoadResources()
 
     ScopedPtr<Camera> camera(new Camera());
     VirtualCoordinatesSystem* vcs = DAVA::UIControlSystem::Instance()->vcs;
-    float32 aspect = (float32)vcs->GetVirtualScreenSize().dy / (float32)vcs->GetVirtualScreenSize().dx;
+    float32 aspect = static_cast<float32>(vcs->GetVirtualScreenSize().dy) / vcs->GetVirtualScreenSize().dx;
     camera->SetupPerspective(70.f, aspect, 0.5f, 2500.f);
     camera->SetLeft(Vector3(1, 0, 0));
     camera->SetUp(Vector3(0, 0, 1.f));
